apps/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
requests==2.32.3
supabase==2.6.0
pydantic>=2,<3

python-jose[cryptography]==3.3.0
brevo-python
email-validator
aiortc==1.9.0
av==12.0.0
numpy==1.26.4
aiofiles==24.1.0
